# openapi.json
{
  "components": {
    "schemas": {
      "Body_submit_fetch_result_api_v1_gateways_jobs__job_id__fetch_result_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_submit_fetch_result_api_v1_gateways_jobs__job_id__fetch_result_post",
        "type": "object"
      },
...
      "GatewayCreate": {
        "properties": {
          "name": {
            "default": "Gateway",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "title": "GatewayCreate",
        "type": "object"
      },
      "GatewayCredentials": {
        "properties": {
          "gateway_id": {
            "format": "uuid",
            "title": "Gateway Id",
            "type": "string"
          },
          "gateway_key": {
            "title": "Gateway Key",
            "type": "string"
          },
          "gateway_online_seconds": {
            "default": 60,
            "title": "Gateway Online Seconds",
            "type": "integer"
          },
          "pairing_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pairing Expires At"
          },
          "pairing_token": {
            "title": "Pairing Token",
            "type": "string"
          },
          "pairing_token_ttl_minutes": {
            "default": 15,
            "title": "Pairing Token Ttl Minutes",
            "type": "integer"
          }
        },
        "required": [
          "gateway_id",
          "pairing_token",
          "gateway_key"
        ],
        "title": "GatewayCredentials",
        "type": "object"
      },
...
      "GatewayJobStatusOut": {
        "properties": {
          "attempts": {
            "default": 0,
            "title": "Attempts",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "library_item_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Library Item Id"
          },
          "payload": {
            "title": "Payload",
            "type": "object"
          },
          "status": {
            "$ref": "#/components/schemas/GatewayJobStatus"
          },
          "type": {
            "$ref": "#/components/schemas/GatewayJobType"
          }
        },
        "required": [
          "job_id",
          "type",
          "payload",
          "status"
        ],
        "title": "GatewayJobStatusOut",
        "type": "object"
      },
...
      "GatewayOut": {
        "properties": {
          "gateway_id": {
            "format": "uuid",
            "title": "Gateway Id",
            "type": "string"
          },
          "gateway_online_seconds": {
            "default": 60,
            "title": "Gateway Online Seconds",
            "type": "integer"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pairing_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pairing Expires At"
          },
          "pairing_token_ttl_minutes": {
            "default": 15,
            "title": "Pairing Token Ttl Minutes",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/PairingStatus"
          }
        },
        "required": [
          "gateway_id",
          "name",
          "status",
          "last_seen_at"
        ],
        "title": "GatewayOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchResults": {
        "description": "Liste JSON nue acceptee depuis le gateway-agent.",
        "items": {
          "$ref": "#/components/schemas/Result"
        },
        "title": "SearchResults",
        "type": "array"
      },
...
    }
  },
  "info": {
    "title": "Ferry Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/gateways": {
      "get": {
        "operationId": "list_gateways_api_v1_gateways_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Dev-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GatewayOut"
                  },
                  "title": "Response List Gateways Api V1 Gateways Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Gateways",
        "tags": [
          "gateways"
        ]
      },
      "post": {
        "operationId": "create_gateway_api_v1_gateways_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Dev-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/GatewayCreate"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayCredentials"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Gateway",
        "tags": [
          "gateways"
        ]
      }
    },
    "/api/v1/gateways/jobs/{job_id}": {
      "get": {
        "operationId": "get_gateway_job_status_api_v1_gateways_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Dev-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayJobStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Gateway Job Status",
        "tags": [
          "gateways"
        ]
      }
    },
    "/api/v1/gateways/jobs/{job_id}/fetch-result": {
      "post": {
        "operationId": "submit_fetch_result_api_v1_gateways_jobs__job_id__fetch_result_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Gateway-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Gateway-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_fetch_result_api_v1_gateways_jobs__job_id__fetch_result_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayFetchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Fetch Result",
        "tags": [
          "gateways"
        ]
      }
    },
    "/api/v1/gateways/jobs/{job_id}/search-results": {
      "post": {
        "operationId": "submit_search_results_api_v1_gateways_jobs__job_id__search_results_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Gateway-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Gateway-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchResults"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayJobAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Search Results",
        "tags": [
          "gateways"
        ]
      }
    },
    "/api/v1/gateways/pair": {
      "post": {
        "operationId": "pair_gateway_api_v1_gateways_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayPair"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayId"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pair Gateway",
        "tags": [
          "gateways"
        ]
      }
    },
    "/api/v1/gateways/poll": {
      "post": {
        "operationId": "poll_gateway_api_v1_gateways_poll_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Gateway-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Gateway-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/GatewayJobOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Poll Gateway Api V1 Gateways Poll Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Poll Gateway",
        "tags": [
          "gateways"
        ]
      }
    },
    "/api/v1/gateways/revoke": {
      "post": {
        "operationId": "revoke_gateway_api_v1_gateways_revoke_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Dev-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayRevoke"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayId"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Gateway",
        "tags": [
          "gateways"
        ]
      }
    },
    "/api/v1/gateways/{gateway_id}": {
      "delete": {
        "operationId": "delete_gateway_api_v1_gateways__gateway_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "gateway_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Gateway Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Dev-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Gateway",
        "tags": [
          "gateways"
        ]
      }
    },
    "/api/v1/gateways/{gateway_id}/jobs": {
      "get": {
        "operationId": "list_gateway_jobs_api_v1_gateways__gateway_id__jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "gateway_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Gateway Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Dev-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GatewayJobStatusOut"
                  },
                  "title": "Response List Gateway Jobs Api V1 Gateways  Gateway Id  Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Gateway Jobs",
        "tags": [
          "gateways"
        ]
      }
    },
    "/api/v1/gateways/{gateway_id}/recreate": {
      "post": {
        "description": "Regenere le code de connexion et la cle d'acces (affichage unique).",
        "operationId": "recreate_gateway_api_v1_gateways__gateway_id__recreate_post",
        "parameters": [
          {
            "in": "path",
            "name": "gateway_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Gateway Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Dev-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayCredentials"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recreate Gateway",
        "tags": [
          "gateways"
        ]
      }
    },
...
  }
}